requirements.txt
aiogram
Pillow
easyocr
numpy